clean up reporter correctly
