+++ packages/app/src/runner/injectBundle.ts
@@ -1,6 +1,3 @@
-import { initializeEventManager } from '.'
-import { initializeMobxStore, useAutStore } from '../store'
-
 export async function injectBundle () {
   const src = '/__cypress/runner/cypress_runner.js'
 
@@ -10,8 +7,6 @@ export async function injectBundle () {
     return
   }
 
-  const response = await window.fetch('/api')
-  const data = await response.json()
   const script = document.createElement('script')
 
   script.src = src
@@ -29,26 +24,6 @@ export async function injectBundle () {
     script.onload = () => {
       // just stick config on window until we figure out how we are
       // going to manage it
-      const config = window.UnifiedRunner.decodeBase64(data.base64Config) as any
-      const autStore = useAutStore()
-
-      // TODO(lachlan): use GraphQL to get the viewport dimensions
-      // once it is more practical to do so
-      // find out if we need to continue managing viewportWidth/viewportHeight in MobX at all.
-      autStore.updateDimensions(config.viewportWidth, config.viewportHeight)
-
-      window.UnifiedRunner.config = config
-
-      // window.UnifiedRunner exists now, since the Webpack bundle with
-      // the UnifiedRunner namespace was injected.
-      initializeEventManager(window.UnifiedRunner)
-
-      window.UnifiedRunner.MobX.runInAction(() => {
-        const store = initializeMobxStore(window.UnifiedRunner.config.testingType)
-
-        store.updateDimensions(config.viewportWidth, config.viewportHeight)
-      })
-
       resolve()
     }
   })
